requirements.txt
beautifulsoup4
aiohttp
async_timeout>=3.0,<4.0
urllib3
requests
lxml
tqdm
flask
python-xbrl
Flask-SQLAlchemy
PyJWT